\name{expand.gamma}
\Rdversion{1.1}
\alias{expand.gamma}

\title{
expand.gamma
}
\description{
Expand a state transition matrix to a number of replicates, i.e. each column is replicated reps times.
}
\usage{
expand.gamma(gammaposs, reps=4, order=TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gammaposs}{State transition matrix}
  \item{reps}{Number of replicates}
  \item{order}{Boolean. Reorder after expansion.}
}
\details{
TODO
}
\value{
Expanded state transition matrix gamma.
}
\author{
Christian Bender
}
%\examples{
%## none
%}

\keyword{ internal }
