\name{mcmc_accept}
\alias{mcmc_accept}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mcmc_accept - accept move or not
}
\description{
Computes acceptance rates for the move as well as for the prior strength parameter lambda.
}
\usage{
mcmc_accept(bestmodel, bettermodels, newlambda, debug=0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bestmodel}{The current model. }
  \item{bettermodels}{The proposed model. }
  \item{newlambda}{The proposed prior strength. }
  \item{debug}{Numeric. The debug level to be used.}
}
\details{
TODO
}
\value{
TODO
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}

\seealso{
\code{\link{mcmc_ddepn}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ internal }

