\name{get.aic}
\Rdversion{1.1}
\alias{get.aic}
\alias{get.bic}

\title{
get.aic, get.bic
}
\description{
Model comparison scores incorporating the complexity of the model.
}
\usage{
get.aic(phi,L)
get.bic(phi,L,n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phi}{Network}
  \item{L}{Likelihood}
  \item{n}{Number of data points.}
}
\details{
TODO
}
\value{
The score.
}
\author{
Christian Bender
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ internal }
