\name{random.gammaprime}
\Rdversion{1.1}
\alias{random.gammaprime}
\title{
random.gammaprime
}
\description{
Sample a random state transition matrix. Usually called internally.
}
\usage{
random.gammaprime(gammax)
}
\arguments{
  \item{gammax}{The possible state vectors.}
}
\details{
TODO
}
\value{
TODO
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}
\seealso{
TODO
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ internal }
