\name{uniquegamma}
\Rdversion{1.1}
\alias{uniquegamma}
\alias{uniquegammaposs}

\title{
uniquegamma
}
\description{
Get rid of multiple equal state vectors in a state transition matrix.
}
\usage{
uniquegamma(gammax)
uniquegammaposs(gammaposs.tmp)
}
\arguments{
  \item{gammax}{}
  \item{gammaposs.tmp}{}
}
\details{
TODO
}
\value{
TODO
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}
\seealso{
TODO
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
\keyword{ internal }
