\name{expand.gamma}
\Rdversion{1.1}
\alias{expand.gamma}

\title{
expand.gamma
}
\description{
Expand a state transition matrix to a number of replicates, i.e. each column is replicated reps times.
}
\usage{
expand.gamma(e.gamma, reps=4, order=TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{e.gamma}{State transition matrix}
  \item{reps}{Number of replicates}
  \item{order}{}
}
\details{
TODO
}
\value{
Expanded state transition matrix gamma.
}
\author{
Christian Bender
}
\examples{
## none
}

\keyword{ internal }
