\name{kegggraph.to.detailed.adjacency}
\Rdversion{1.1}
\alias{kegggraph.to.detailed.adjacency}

\title{
kegggraph.to.detailed.adjacency
}
\description{
Converts a keggraph object to an adjacency matrix containing inhibitions and activations.
}
\usage{
kegggraph.to.detailed.adjacency(gR)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gR}{A KEGGgraph object.}
}
\details{
Internal function. Shouldn't be called directly.
}
\value{
An adjacency matrix.
}
\author{
Christian Bender
%%  ~~who you are~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}

\keyword{ internal }
%\keyword{ ~matrix }% __ONLY ONE__ keyword per line
