\name{netga}
\Rdversion{1.1}
\alias{netga}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
netga - Perform network search using a genetic algorithm.
}
\description{
Uses a genetic algorithm for network search. Combines selection of best models,
crossing over of nodes and connected edges between pairs of nodes and mutation
of edges to sample the search space.
}
\usage{
netga(datx, stimuli, P=NULL, maxiterations=1000, p=100,
		q=0.3, m=0.8, maxiter=30, multicores=FALSE, usebics=FALSE, cores=2,
		lambda=NULL, B=NULL,
		Z=NULL, scorefile="score.pdf")
}
\arguments{
  \item{datx}{
	The data matrix. See \code{\link{ddepn}}
}
  \item{stimuli}{
	The list of stimuli. See \code{\link{ddepn}}
}
  \item{P}{
	A list containing a population of network models. If NULL, an initial population
	of networks is sampled including the empty and fully connected network.
}
  \item{maxiterations}{
	Integer, maximal number of generations.
}
  \item{p}{
	The size of the population. Used only if P is not provided.
}
  \item{q}{
	Double in [0;1]. The rate of selection or crossing over. At most \eqn{(1-q) \cdot |P|}
	nodes are selected to persist into the next generation, but only if the fitness score more optimal
	than the median of fitnessess in the population. If less than \eqn{(1-q)*|P|} fulfill this criterion,
	the number \eqn{q*|P|} of crossing overs is increased accordingly.
}
  \item{m}{
	Double in [0;1]. The mutation rate - \eqn{m \cdot |P|} networks out of the Population 
	are mutated \emph{after} selection and crossing over. Mutations happen for edges and 
	change the type of the edge randomly.
}
  \item{maxiter}{
	Maximum number of iterations during the HMM.
}
  \item{multicores}{
	Use multiple processors.
}
  \item{usebics}{
    Use BIC statistic for model selection.
}
  \item{cores}{
    Number of cores to use in case of \code{multicores}=TRUE
}
  \item{lambda}{
  	The Prior influence strength.
  }
  \item{B}{
  	The Prior information matrix.
  }
  \item{Z}{
  	The normalisation factor for the prior distribution.
  }
  \item{scorefile}{
  	Name of pdf-file where intermediate traces of the BIC, the differences of BICs in two succeeding 
  	generations and a comparison of optimum and average fitness are drawn.
  }
}
\details{
TODO
}
\value{
A list containing the model objects for each candidate network in the population. Each 
entry in the list contains the following elements:
  \item{phi}{The inferred network}
  \item{L}{Likelihood}
  \item{aic}{Akaikes Information Criterion}
  \item{bic}{Bayesian Information Criterion}
  \item{posterior}{Posterior probability}
  \item{dat}{The data}
  \item{theta}{The parameter matrix for the gaussians.}
  \item{gamma}{The state transition matrix.}
  \item{tps}{Timepoints}
  \item{stimuli}{List of stimuli}
  \item{reps}{Number of replicates}
  \item{maxiter}{Maximum iterations to performed.}
  \item{TSA}{deprecated}
  \item{Tt}{deprecated}  
  \item{lastmove}{Type of the last change that was performed.}
  \item{coords}{Position in the network where the last change was performed.}
  \item{lambda}{Prior strength parameter.}
  \item{B}{Prior matrix}
  \item{Z}{Prior normalisation factor.}
  \item{gammaposs}{The theoretical state transition matrix, as generated by the effect propagation.}
  \item{iterations}{The number of iterations performed in the GA.}
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO
}
\examples{
\dontrun{
## load package
library(ddepn)

## sample a network
n <- 6
signet <- signalnetwork(n=n, nstim=2, cstim=0, prop.inh=0.2)
phit <- signet$phi
stimuli <- signet$stimuli

## sample data
dataset <- makedata(phit, stimuli, mu.bg=1200, sd.bg=400, mu.signal.a=2000, sd.signal.a=1000)

## now the GA
ret <- netga(dataset$dat, stimuli, P=NULL, maxiterations=50, p=10,
    q=0.3, m=0.8, maxiter=100, multicores=FALSE, usebics=TRUE, cores=1,
    lambda=NULL, B=NULL, Z=NULL, scorefile=NULL)


}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
