\name{uniquegamma}
\Rdversion{1.1}
\alias{uniquegamma}
\alias{uniquegammaposs}

\title{
uniquegamma
}
\description{
Get rid of multiple equal state vectors in a state transition matrix.
}
\usage{
uniquegamma(gammax)
uniquegammaposs(gammaposs.tmp)
}
\arguments{
  \item{gammax}{Optimised state matrix.}
  \item{gammaposs.tmp}{Theoretical state matrix.}
}
\details{
TODO
}
\value{
TODO
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}
\seealso{
TODO
}
\examples{
##TODO
}
\keyword{ internal }
